Week_21/Day_2/RAG_technique and Evaluation/requirements.txt
chromadb==1.3.7
langchain_chroma==1.1.0
langchain_classic==1.0.0
langchain_community==0.4.1
langchain_core==1.2.0
langchain_experimental==0.4.1
langchain_openai==1.1.3
langchain_text_splitters==1.1.0
pandas==2.3.3
python-dotenv==1.2.1
ragas==0.4.1
numpy==1.26.4
nltk==3.9.2
sentence-transformers==5.2.0